trustbench_groq_starter/pyproject.toml
[project]
name = "trustbench-core"
version = "0.1.0"
description = "Groq-first agent evaluation scaffold"
requires-python = ">=3.10"
dependencies = [
  "pyyaml",
  "pandas",
  "rich",
  "jinja2",
  "groq>=0.9.0",
  "streamlit"
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "black"]

[tool.ruff]
line-length = 100
